Added screen scaling
Application will now scale text and images according to your screen resolution.

It compares the screen to a 1920x1080 screen and determines factor to scale by. All text, the graphs, and the base 64 encoded images on the main screen and title bar and scaled by this factor to help ensure everything is visible on screen.

Originally I wanted to use the scaling parameter in PySimpleGUI, however there is a bug with the latest version of Matplotlib that causes the graphs to shrink upon repeated redraws. This solution will be used instead.