Updated Read Me
Seeing as my internship ends in a week, I am marking the project as completed. My last updates have now been posted to the Read Me and the repository.
I have also written documentation on usage, repairing, and installing the unit for future years when I am no longer with the company.

Overall, everyone has been very pleased with the result and I hope to see the unit run for many years to come.