Working on dark mode
Something I have been trying to figure out for a while, but couldn't because of my lack of knowledge of Tkinter.
Thanks to definite-d (https://github.com/definite-d/PSG_Reskinner/blob/main/src/__main__.py) I found out how this can be accomplished.

Still in the early phases, as the titlebar and some other elements are changing despite multiple different attempts to prevent it. Will continue working on it.